refactor: testFixture 패키지로 인한 로직 변경
